# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UAC API Documentation",
    "version": "1.0.0",
    "description": "用户认证和授权系统 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "角色ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "role_name": {
            "type": "string",
            "description": "角色名称",
            "example": "管理员"
          },
          "code": {
            "type": "string",
            "description": "角色编码",
            "example": "ADMIN"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员角色"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "ARCHIVED"
            ],
            "description": "角色状态",
            "example": "ACTIVE"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间",
            "example": "2024-03-21T10:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间",
            "example": "2024-03-21T10:00:00.000Z"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "删除时间",
            "example": null
          },
          "permissions": {
            "type": "array",
            "description": "角色权限列表",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "RoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/Role"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/dev/login": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "开发环境专用登录接口",
        "description": "仅限开发环境使用的登录接口，无需验证码",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "登录成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "JWT token",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        },
                        "refresh_token": {
                          "type": "string",
                          "description": "刷新令牌",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        },
                        "expires_in": {
                          "type": "string",
                          "description": "令牌过期时间",
                          "example": "2h"
                        },
                        "user_id": {
                          "type": "string",
                          "description": "用户ID",
                          "example": "550e8400-e29b-41d4-a716-446655440000"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "user_id": {
                              "type": "string"
                            },
                            "username": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "roles": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "role_id": {
                                    "type": "string"
                                  },
                                  "role_name": {
                                    "type": "string"
                                  },
                                  "code": {
                                    "type": "string"
                                  },
                                  "permissions": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "permission_id": {
                                          "type": "string"
                                        },
                                        "permission_name": {
                                          "type": "string"
                                        },
                                        "code": {
                                          "type": "string"
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "用户名或密码错误"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          },
          "403": {
            "description": "非开发环境",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "用户已被禁用"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "服务器内部错误"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/permissions/{permission_id}/roles": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "summary": "分配角色权限",
        "description": "为指定角色分配权限",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "权限ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role_ids"
                ],
                "properties": {
                  "role_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "角色ID列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分配成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "权限分配成功"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "404": {
            "description": "权限不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
    "/api/v1/roles/{role_id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "获取角色详情",
        "description": "获取指定角色的详细信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "角色不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "更新角色信息",
        "description": "更新指定角色的信息",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "角色ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role_name": {
                    "type": "string",
                    "description": "角色名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "角色描述"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "角色不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "删除角色",
        "description": "软删除指定角色",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "角色删除成功"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "角色不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
    "/api/v1/roles/check-permission": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "检查用户权限",
        "description": "检查当前用户是否拥有指定权限",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "permission_code",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "权限编码"
          }
        ],
        "responses": {
          "200": {
            "description": "检查成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "has_permission": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  },
...
}